Update SliceAndRotate.py
This function can be called directly on X_all and Y_all as returned by Zack's code.  It will return an N-dimensional array of the same shape as the argument passed to the function, but the length of the 0th dimension will be 32x longer (there are 4 quadrants and 8 distinct transformations on each quadrant).  The dytpe of the array returned is float64.  The array returned can be converted back to the original datatypes as follows:

X_all = SliceAndRotate(X_all).astype('uint8')
Y_all = SliceAndRotate(Y_all) .astype('bool') OR Y_all_enhanced = SliceAndRotate(Y_all) > some_arbitrary_threshold

It should be possible to drop these lines in right after X_all, Y_all = read_train_data(), and fed into the U-Net.